requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1

# Database drivers
# PostgreSQL
psycopg2-binary==2.9.9
# SQLite is built-in to Python

# Database migrations
alembic==1.13.0
Flask-Migrate==4.0.5

# Environment management
python-dotenv==1.0.0

# Machine Learning (existing)
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
pickle-mixin==1.0.2

# Image processing (for future multimodal support)
Pillow==10.1.0
opencv-python==4.8.1.78

# Audio processing (for future multimodal support)
librosa==0.10.1
soundfile==0.12.1

# Deep Learning (PyTorch - for future multimodal models)
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

# Cloud Storage
# AWS S3
boto3==1.34.10
# Azure Blob Storage
azure-storage-blob==12.19.0

# Async task processing
celery==5.3.4
redis==5.0.1

# HTTP and API
requests==2.31.0

# File handling
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Production server
gunicorn==21.2.0
